Use envvars for all configuration
Previously dev mode, debug mode, and bind ports were cmd options.

This made syncing the port automatically with the health check impossible. This is more consistent anyway, all config is now done through envvars.
