+++ cmd/dims/serve.go
@@ -1,0 +1,48 @@
+// Copyright 2025 Jeremy Collins. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package main
+
+import (
+	"log/slog"
+	"net/http"
+	"os"
+
+	"github.com/beetlebugorg/go-dims/internal/dims/core"
+	"github.com/beetlebugorg/go-dims/pkg/dims"
+)
+
+type ServeCmd struct {
+}
+
+func (s *ServeCmd) Run() error {
+	config := core.ReadConfig()
+
+	var opts *slog.HandlerOptions
+	if config.DebugMode {
+		opts = &slog.HandlerOptions{
+			Level: slog.LevelDebug,
+		}
+	}
+
+	logger := slog.New(slog.NewJSONHandler(os.Stdout, opts))
+	slog.SetDefault(logger)
+
+	err := http.ListenAndServe(config.BindAddress, dims.NewHandler(config))
+	if err != nil {
+		slog.Error("Server failed.", "error", err)
+		return err
+	}
+	return nil
+}
