+++ internal/dims/status.go
@@ -20,7 +20,7 @@ import (
 	"github.com/beetlebugorg/go-dims/internal/dims/core"
 )
 
-func HandleDimsStatus(config core.EnvironmentConfig, debug bool, dev bool, w http.ResponseWriter, r *http.Request) {
+func HandleDimsStatus(config core.Config, w http.ResponseWriter, r *http.Request) {
 	w.WriteHeader(200)
 	w.Write([]byte("ALIVE"))
 }
